Consolidate into AoiSerializer
Where all the logic for translating WKAoI to AoI was
already collected.
